pyproject.toml
[tool.poetry]
name = "netdevops-survey"
version = "0.1.0"
description = ""
authors = ["Damien Garros <dgarros@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.7"
sqlalchemy = "^1.3.20"
pyyaml = "^5.3.1"
pygal = "^2.4.0"
pandas = "^1.1.4"
cairosvg = "^2.5.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"


[tool.poetry.scripts]
netdevops-survey = 'netdevops_survey.cli:main'

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
        /(
          \.git
        | \.tox
        | \.venv
        | env/
        | _build
        | build
        | dist
        )/
    '''

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long,
    bad-continuation,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """
    
[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
